+++ libs/preprocessor/parser.h
@@ -1,5 +1,5 @@
 /*
- *	Copyright 2020 Andrey Terekhov, Egor Anikin
+ *	Copyright 2018 Andrey Terekhov, Egor Anikin
  *
  *	Licensed under the Apache License, Version 2.0 (the "License");
  *	you may not use this file except in compliance with the License.
@@ -18,12 +18,12 @@
 
 #include "environment.h"
 
-
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-int if_implementation(environment *const env);
+
+int preprocess_scan(environment *const env);
 
 #ifdef __cplusplus
 } /* extern "C" */
